@@ -32,7 +32,7 @@ dependencies:
   json_serializable: ^6.9.0
   google_maps_flutter: ^2.10.0
   flutter_google_maps_webservices: ^1.1.1
-  google_places_flutter: ^2.0.6
+  google_places_flutter: ^2.1.0
   location: ^8.0.0
   flutter_polyline_points: ^2.1.0
   geocoding: ^3.0.0
